+++ lib/graphql/execution/multiplex.rb
@@ -91,7 +91,7 @@ module GraphQL
             # Get the Query::Result, not the Hash
             query.result
           end
-        rescue StandardError
+        rescue Exception
           # Assign values here so that the query's `@executed` becomes true
           queries.map { |q| q.result_values ||= {} }
           raise