+++ lib/graphql/schema/resolver.rb
@@ -23,6 +23,8 @@ module GraphQL
       # Really we only need description from here, but:
       extend Schema::Member::BaseDSLMethods
       extend GraphQL::Schema::Member::HasArguments
+      include Schema::Member::HasPath
+      extend Schema::Member::HasPath
 
       # @param object [Object] the initialize object, pass to {Query.initialize} as `root_value`
       # @param context [GraphQL::Query::Context]
@@ -72,7 +74,6 @@ module GraphQL
               else
                 authorized?
               end
-              authorized?(loaded_args)
               context.schema.after_lazy(authorized_val) do |(authorized_result, early_return)|
                 # If the `authorized?` returned two values, `false, early_return`,
                 # then use the early return value instead of continuing
@@ -323,13 +324,32 @@ module GraphQL
         # @see {GraphQL::Schema::Argument#initialize} for the signature
         def argument(name, type, *rest, loads: nil, **kwargs, &block)
           if loads
-            arg_keyword = name.to_s.sub(/_id$/, "").to_sym
-            kwargs[:as] = arg_keyword
-            own_arguments_loads_as_type[arg_keyword] = loads
+            name_as_string = name.to_s
+
+            inferred_arg_name = case name_as_string
+            when /_id$/
+              name_as_string.sub(/_id$/, "").to_sym
+            when /_ids$/
+              name_as_string.sub(/_ids$/, "")
+                .sub(/([^s])$/, "\\1s")
+                .to_sym
+            else
+              name
+            end
+
+            kwargs[:as] ||= inferred_arg_name
+            own_arguments_loads_as_type[kwargs[:as]] = loads
           end
+
           arg_defn = super(name, type, *rest, **kwargs, &block)
 
-          if loads
+          if loads && arg_defn.type.list?
+            class_eval <<-RUBY, __FILE__, __LINE__ + 1
+            def load_#{arg_defn.keyword}(values)
+              GraphQL::Execution::Lazy.all(values.map { |value| load_application_object(:#{arg_defn.keyword}, value) })
+            end
+            RUBY
+          elsif loads
             class_eval <<-RUBY, __FILE__, __LINE__ + 1
             def load_#{arg_defn.keyword}(value)
               load_application_object(:#{arg_defn.keyword}, value)